+++ routes/index.js
@@ -1,13 +1,13 @@
 const path = require("path");
 const router = require("express").Router();
-const apiRoutes = require("./api");
+const apiRoutes = require("./api/api");
 
 // API Routes
 router.use("/api", apiRoutes);
 
 // If no API routes are hit, send the React app
-router.use((req, res) =>
-  res.sendFile(path.join(__dirname, "../client/build/index.html"))
-);
+// router.use((req, res) =>
+//   res.sendFile(path.join(__dirname, "../client/build/index.html"))
+// );
 
 module.exports = router;
