+++ server.js
@@ -2,6 +2,7 @@ const express = require("express");
 const mongoose = require("mongoose");
 const crypto = require("crypto");
 const path = require("path");
+const Grid = require("gridfs-stream");
 const routes = require("./routes");
 const app = express();
 const PORT = process.env.PORT || 3001;
@@ -17,41 +18,29 @@ if (process.env.NODE_ENV === "production") {
 // Add routes, both API and view
 app.use(routes);
 // Connect to the Mongo DB
-let conn = mongoose.connect(
-  process.env.MONGODB_URI || "mongodb://localhost/pocIT",
-  {
-    useCreateIndex: true,
-    useNewUrlParser: true,
-  }
-);
+mongoose.connect(process.env.MONGODB_URI || "mongodb://localhost/pocIT", {
+  useCreateIndex: true,
+  useNewUrlParser: true,
+});
 
 let gfs;
 
-conn.once("open", () => {
-  gfs = Grid(conn.db, mongoose.mongo);
+const db = require("./models");
+
+// db.Patient.create({
+//   isProvider: false,
+//   email: "test@test.com",
+//   password: "testing",
+//   name: "mr test",
+//   dob: 1,
+//   height: 6,
+//   weight: 180,
+// });
+
+mongoose.connection.once("open", () => {
+  gfs = Grid(mongoose.connection.db, mongoose.mongo);
   gfs.collection("uploads");
 });
-
-const storage = new GridFsStorage({
-  url: "mongodb://localhost/pocIT",
-  file: (req, file) => {
-    return new Promise((resolve, reject) => {
-      crypto.randomBytes(16, (err, buf) => {
-        if (err) {
-          return reject(err);
-        }
-        const filename = buf.toString("hex") + path.extname(file.originalname);
-        const fileInfo = {
-          filename: filename,
-          bucketName: "uploads",
-        };
-        resolve(fileInfo);
-      });
-    });
-  },
-});
-
-const upload = multer({ storage });
 
 // Start the API server
 app.listen(PORT, () =>
